requirements_simple.txt
streamlit>=1.28.0
pandas>=1.5.0
PyMuPDF>=1.20.0
python-docx>=0.8.0
